refactored the main.rb file
