+++ classes/student.rb
@@ -1,4 +1,4 @@
-require './person'
+require_relative './person'
 
 class Student < Person
   attr_reader :classroom
